task-analyze/backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==3.14.0
